review(cerebrum): harden JSON metadata lift + drop service ref from AC
CodeRabbit follow-ups on #2607:

* Move JSON detection into a dedicated `ingest-json.ts` so the MCP handler
  file stays under the lint cap.
* Block prototype-pollution keys (`__proto__`, `prototype`, `constructor`)
  when lifting JSON into `customFields`, and create the bag with
  `Object.create(null)` so the engram frontmatter can't be tampered with by
  agent-supplied JSON.
* `pickStringArray` now trims the values it returns, not just the values it
  filters by — caller-supplied `[" work.sprint ", " personal "]` now lands
  as `["work.sprint", "personal"]`.
* Drop the `IngestService.quickCapture` reference from PRD US-02 AC #2 so
  the acceptance criteria stay behaviour-only.

Skipped: CodeRabbit's suggestion to also parse JSON primitives (e.g.
`"hello"`, `42`) — the current `{` / `[` gate is deliberate. Wrapping plain
Markdown that happens to be JSON-parseable in a code block produces worse
output than the current behaviour where only structured objects and arrays
take the JSON path.
